requirements.txt
streamlit==1.32.2
python-docx==1.1.0
PyYAML==6.0.1
bcrypt==4.1.2
pymongo==4.6.3
requests==2.31.0
PyPDF2==3.0.1
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
